Create Problem3
Entering a number in a text box on AIO determines the number of blinks of on-board USR3 LED. In general, this message/number received by the BBB can be used to determine what to do next; e.g., invoke another program, save the messages in a file, etc. Must use Molloy’s LED class.