Create Problem5
Virtual toggle button on AIO that controls AC appliances such as TVs, desk lamps, electric heaters, air conditioners, etc. A GPIO pin on the BBB controls the appliances (can use an LED that is interfaced to a BBB pin to indicate that the appliance is ON). Must use Molloy’s GPIO class.